Consolidate strategy types and remove unused imports
